fix: change the location of definition of type related configurations
