+++ packages/vscode-wdio-types/src/utils.ts
@@ -1,11 +1,5 @@
 import type { LOG_LEVEL } from '@vscode-wdio/constants'
-
-export interface WebdriverIOConfig {
-    nodeExecutable: string | undefined
-    configFilePattern: string[]
-    showOutput: boolean
-    logLevel: string
-}
+export type { WebdriverIOConfig } from '@vscode-wdio/constants'
 
 export type WdioLogLevel = 'trace' | 'debug' | 'info' | 'warn' | 'error' | 'silent'
